requirements.txt
datasets==3.6.0
duckdb==1.2.2
evaluate==0.4.3
hydra-core==1.3.2
numpy==2.2.5
pandas==2.2.3
torch==2.7.0
transformers==4.51.3
